Общая логика компонентов точки события вынесена в BaseEvent
